refactor: 회원가입 함수 분리
