+++ src/components/pages/JoinPage.tsx
@@ -4,6 +4,7 @@ import styled from "@emotion/styled";
 import GotoLogin from "../join/GotoLogin";
 import { useState } from "react";
 import { useNavigate } from "react-router-dom";
+import { handleSignup } from "../../services/handleSignup";
 
 const JoinContainerStyle = styled.div`
   display: flex;
@@ -37,42 +38,7 @@ const JoinPage = () => {
   const [newPw, setNewPw] = useState<string>("");
   const [checkNewPw, setCheckNewPw] = useState<string>("");
   const [newAlias, setNewAlias] = useState<string>("");
-
-  const handleSignup = async () => {
-    const navigate = useNavigate();
-    if (newPw !== checkNewPw) {
-      alert("비밀번호가 일치하지 않습니다.");
-      return;
-    }
-    const signupData = {
-      loginId: newId,
-      username: newAlias,
-      password: newPw,
-    };
-
-    try {
-      const response = await fetch(
-        "http://13.125.208.182:8080/v1/members/join",
-        {
-          method: "POST",
-          headers: { "Content-Type": "application/json" },
-          body: JSON.stringify(signupData),
-          credentials: "include",
-        }
-      );
-      if (response.ok) {
-        alert("회원가입에 성공했습니다!");
-        navigate("/login");
-      } else {
-        const errorData = await response.json();
-        alert(`회원가입 실패: ${errorData.message}`);
-        console.log(`Error status : ${response.status}`);
-      }
-    } catch (error) {
-      console.log("회원가입 요청 중 오류 발생:", error);
-      alert("회원가입 요청 중 문제가 발생했습니다.");
-    }
-  };
+  const navigator = useNavigate();
 
   return (
     <>
@@ -89,7 +55,11 @@ const JoinPage = () => {
             checkNewPw={checkNewPw}
             setCheckNewPw={setCheckNewPw}
           />
-          <Button handleSignup={handleSignup} />
+          <Button
+            handleSignup={() =>
+              handleSignup(newId, newAlias, newPw, checkNewPw, navigator)
+            }
+          />
           <GotoLogin />
         </JoinWperStyle>
       </JoinContainerStyle>
